Enhance Production Output and Inventory Slip Functionality
- Added unique constraint to MaterialOutputMapping to prevent duplicate mappings.
- Updated ProductionOutput model to use decimal for Finished and Defected properties.
- Refactored services and DTOs to ensure consistent handling of decimal values for quantities.
- Improved front-end components for better user experience, including validation and input handling for quantities.
- Streamlined mapping logic in InventorySlipService to avoid duplicates and ensure accurate processing of raw materials and semi-finished products.

+++ SEP490/Modules/ChemicalExportModule/DTO/ChemicalExportDto.cs
@@ -62,8 +62,8 @@ namespace SEP490.Modules.ChemicalExportModule.DTO
         public string? ProductName { get; set; }
         public string? UOM { get; set; }
         public decimal? Amount { get; set; }
-        public int? Finished { get; set; }
-        public int? Defected { get; set; }
+        public decimal? Finished { get; set; }
+        public decimal? Defected { get; set; }
     }
 
     public class ProductionMaterialDto
